+++ app/Controllers/Http/Business/Task.ts
@@ -1,6 +1,6 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { Task as TaskValidator } from 'App/Validators/Business'
-import { Task, User } from 'App/Models'
+import { Task } from 'App/Models'
 
 export default class TasksController {
   public async index({}: HttpContextContract) {
@@ -9,34 +9,43 @@ export default class TasksController {
   }
 
   public async store({ request, auth }: HttpContextContract) {
-    const data = await request.validate(TaskValidator)
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    const { name_task, description_task, photo, state } = await request.validate(TaskValidator)
     const user = await auth.authenticate()
-    data.user_sended_id = Number(user.id)
-    const task = await Task.create(data)
+    const task = await Task.create({
+      name_task,
+      description_task,
+      photo,
+      state,
+      user_sended_id: user.id,
+    })
+    await task.load('user_send')
     return task
   }
 
   public async show({ response, params }: HttpContextContract) {
-    const task = await Task.findOrFail(Number(params.id))
+    const task = await Task.find(Number(params.id))
 
     if (!task) return response.notFound({ error: 'task not found' })
     if (task.state === 'finished') return response.unauthorized({ error: 'this task is closed' })
-    const user = await User.findOrFail(Number(task.id))
-    return {
-      task,
-      user,
-    }
+    await task.load('user_send')
+    return task
   }
 
-  public async update({ request, params }: HttpContextContract) {
-    const data = request.only(['state', 'name_task', 'description_task'])
-    const task = await Task.findOrFail(Number(params.id))
+  public async update({ request, params, response }: HttpContextContract) {
+    const data = request.only(['name_task', 'description_task', 'photo', 'state'])
+
+    const task = await Task.find(Number(params.id))
+    if (!task) return response.notFound({ message: 'not found' })
     task.merge(data)
     await task.save()
+    await task.load('user_send')
+    return task
   }
 
-  public async destroy({ params }: HttpContextContract) {
-    const task = await Task.findOrFail(Number(params.id))
+  public async destroy({ params, response }: HttpContextContract) {
+    const task = await Task.find(Number(params.id))
+    if (!task) return response.notFound({ message: 'not found' })
     await task.delete()
   }
 }
